requirements.txt
black
eth-brownie
marshmallow-dataclass
pandas
pre-commit
py-etherscan-api
pycoingecko
pytz
redis
requests
seaborn
web3
